requirements.txt
langchain==0.2.1
langchain-core==0.2.1
langchain-community==0.2.1
unstructured[pdf,docx]
faiss-cpu
pymongo
openai
python-dotenv
streamlit
